pyproject.toml
[project]
name = "terifai"
version = "0.1.5"
description = "Conversation AI that clones your voice."
authors = [
    { name = "Aman Ibrahim", email = "amanmibra@gmail.com" }
]
dependencies = [
    "pipecat-ai[daily,deepgram,cartesia,openai,vad,silero,xtts]>=0.0.39",
    "async-timeout>=4.0.3",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.5",
    "modal>=0.63.41",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.rye.scripts]
build_docker = "./scripts/build_docker.sh"
image = "./scripts/run_docker.sh"
fly_deploy = "./scripts/fly_deploy.sh"
bot = "python -m backend.bot --default --prompt=default"
server = "python -m backend.server --reload"
